# swagger.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"security":[{"bearerAuth":[]}],"paths":{"/register":{"post":{"tags":["auth-controller"],"operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}},"/logout":{"post":{"tags":["auth-controller"],"operationId":"logout","parameters":[{"name":"refresh","in":"cookie","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/login":{"post":{"tags":["auth-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}},"/api/v1/ratings":{"get":{"tags":["rating-controller"],"operationId":"getRatings","parameters":[{"name":"productId","in":"query","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":5}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PaginationResponseRatingDto"}}}}}},"post":{"tags":["rating-controller"],"operationId":"createRating","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRatingDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Rating"}}}}}}},"/api/v1/products":{"get":{"tags":["product-controller"],"operationId":"getProductsUserView","parameters":[{"name":"productFilter","in":"query","required":true,"schema":{"$ref":"#/components/schemas/ProductFilterDto"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PaginationResponseProductCardDto"}}}}}},"post":{"tags":["product-controller"],"operationId":"createProduct","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/CreateProductDto"}}}},"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductDetailsDto"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"403":{"description":"Authorization failed","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}},"delete":{"tags":["product-controller"],"operationId":"deleteProducts","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductIdsDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/products/{id}/images":{"post":{"tags":["product-controller"],"operationId":"addImage","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/FileImageDto"}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/categories":{"get":{"tags":["category-controller"],"operationId":"getCategoriesUserView","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CategoryDto"}}}}}}},"post":{"tags":["category-controller"],"operationId":"createCategory","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/CreateCategoryDto"}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}},"delete":{"tags":["category-controller"],"operationId":"deleteCategory","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/users/{id}/is-active":{"patch":{"tags":["user-controller"],"operationId":"updateUserActiveStatus","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserIsActiveDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/products/{id}":{"get":{"tags":["product-controller"],"operationId":"getProductDetailInformation","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductDetailsDto"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"Can not found product","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}},"patch":{"tags":["product-controller"],"operationId":"updateProductDetails","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProductDetailsDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/products/{id}/thumbnail":{"patch":{"tags":["product-controller"],"operationId":"updateProductThumbnail","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/FileImageDto"}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/categories/{id}":{"get":{"tags":["category-controller"],"operationId":"getCategoryDetails","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CategoryDetailsDto"}}}}}},"patch":{"tags":["category-controller"],"operationId":"updateCategoryDetails","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCategoryDetailsDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/categories/{id}/thumbnail":{"patch":{"tags":["category-controller"],"operationId":"updateCategoryThumbnail","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"fileImageDto","in":"query","required":true,"schema":{"$ref":"#/components/schemas/FileImageDto"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/test":{"get":{"tags":["auth-controller"],"operationId":"test","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/refresh":{"get":{"tags":["auth-controller"],"operationId":"refreshToken","parameters":[{"name":"refresh","in":"cookie","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}},"/api/v1/users/{id}":{"get":{"tags":["user-controller"],"operationId":"getUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDetailsDto"}}}}}}},"/api/v1/admin/users":{"get":{"tags":["admin-controller"],"operationId":"getUsersAdminView","parameters":[{"name":"userFilter","in":"query","required":true,"schema":{"$ref":"#/components/schemas/UserFilterDto"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PaginationResponseUserInAdminDto"}}}}}}},"/api/v1/admin/products":{"get":{"tags":["admin-controller"],"operationId":"getProductsAdminView","parameters":[{"name":"productFilter","in":"query","required":true,"schema":{"$ref":"#/components/schemas/ProductFilterDto"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PaginationResponseProductCardInAdminDto"}}}}}}},"/api/v1/admin/categories":{"get":{"tags":["admin-controller"],"operationId":"getCategoriesAdminView","parameters":[{"name":"categoryFilter","in":"query","required":true,"schema":{"$ref":"#/components/schemas/CategoryFilterDto"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PaginationResponseCategoryInAdminDto"}}}}}}},"/api/v1/products/{product-id}/images/{img-id}":{"delete":{"tags":["product-controller"],"operationId":"deleteProductImages","parameters":[{"name":"product-id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"img-id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}}},"components":{"schemas":{"RegisterRequest":{"type":"object","properties":{"email":{"type":"string"},"password":{"maxLength":2147483647,"minLength":8,"type":"string"},"firstname":{"maxLength":2147483647,"minLength":100,"type":"string"},"lastname":{"maxLength":2147483647,"minLength":100,"type":"string"}}},"AuthResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"username":{"type":"string"},"roles":{"type":"array","items":{"type":"string"}},"accessToken":{"type":"string"}}},"LoginRequest":{"type":"object","properties":{"email":{"type":"string"},"password":{"maxLength":2147483647,"minLength":8,"type":"string"}}},"CreateRatingDto":{"required":["productId"],"type":"object","properties":{"productId":{"type":"integer","format":"int64"},"star":{"maximum":5,"minimum":1,"type":"integer","format":"int32"},"comment":{"type":"string"}}},"Rating":{"type":"object","properties":{"createdBy":{"type":"string"},"createdOn":{"type":"string","format":"date-time"},"lastModifiedBy":{"type":"string"},"lastModifiedOn":{"type":"string","format":"date-time"},"id":{"type":"integer","format":"int64"},"comment":{"type":"string"},"star":{"type":"integer","format":"int32"},"productId":{"type":"integer","format":"int64"}}},"CreateProductDto":{"required":["categoryId","description","isFeatured","name","thumbnailImg"],"type":"object","properties":{"name":{"type":"string"},"price":{"minimum":0,"type":"integer","format":"int64"},"description":{"type":"string"},"thumbnailImg":{"type":"string","format":"binary"},"isFeatured":{"type":"boolean"},"quantity":{"minimum":0,"type":"integer","format":"int32"},"categoryId":{"type":"integer","format":"int64"},"images":{"type":"array","items":{"type":"string","format":"binary"}}}},"CategoryDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"description":{"type":"string"},"thumbnailImg":{"type":"string"}}},"ImageDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"url":{"type":"string"},"type":{"type":"string"}}},"ProductDetailsDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"price":{"type":"integer","format":"int64"},"rating":{"type":"number","format":"double"},"description":{"type":"string"},"thumbnailImg":{"type":"string"},"isFeatured":{"type":"boolean"},"quantity":{"type":"integer","format":"int32"},"categories":{"type":"array","items":{"$ref":"#/components/schemas/CategoryDto"}},"images":{"type":"array","items":{"$ref":"#/components/schemas/ImageDto"}}}},"ErrorResponse":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"message":{"type":"string"}}},"FileImageDto":{"type":"object","properties":{"image":{"type":"string","format":"binary"}}},"CreateCategoryDto":{"required":["description","name","thumbnailImg"],"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"},"thumbnailImg":{"type":"string","format":"binary"}}},"UpdateUserIsActiveDto":{"type":"object","properties":{"isActive":{"type":"boolean"}}},"UpdateProductDetailsDto":{"type":"object","properties":{"name":{"type":"string"},"price":{"type":"integer","format":"int64"},"description":{"type":"string"},"isFeatured":{"type":"boolean"},"quantity":{"type":"integer","format":"int32"},"categoryId":{"type":"integer","format":"int64"}}},"UpdateCategoryDetailsDto":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"}}},"UserDetailsDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"username":{"type":"string"},"firstname":{"type":"string"},"lastname":{"type":"string"},"roles":{"type":"array","items":{"type":"string"}},"isActive":{"type":"boolean"}}},"OwnerDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"firstname":{"type":"string"},"lastname":{"type":"string"}}},"PaginationResponseRatingDto":{"type":"object","properties":{"page":{"type":"integer","description":"Current page","format":"int32"},"total":{"type":"integer","description":"Total elements","format":"int64"},"totalPages":{"type":"integer","description":"Total pages","format":"int32"},"data":{"type":"array","items":{"$ref":"#/components/schemas/RatingDto"}}}},"RatingDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"comment":{"type":"string"},"star":{"type":"integer","format":"int32"},"createdOn":{"type":"string","format":"date-time"},"user":{"$ref":"#/components/schemas/OwnerDto"}}},"ProductFilterDto":{"type":"object","properties":{"name":{"type":"string"},"page":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"minPrice":{"type":"integer","format":"int64"},"maxPrice":{"type":"integer","format":"int64"},"isFeatured":{"type":"boolean"},"sortBy":{"type":"string"},"sortOrder":{"type":"string"},"categoryIds":{"type":"array","items":{"type":"integer","format":"int64"}},"minPriceLessThanMaxPrice":{"type":"boolean"}}},"PaginationResponseProductCardDto":{"type":"object","properties":{"page":{"type":"integer","description":"Current page","format":"int32"},"total":{"type":"integer","description":"Total elements","format":"int64"},"totalPages":{"type":"integer","description":"Total pages","format":"int32"},"data":{"type":"array","items":{"$ref":"#/components/schemas/ProductCardDto"}}}},"ProductCardDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"thumbnailImg":{"type":"string"},"price":{"type":"integer","format":"int64"},"rating":{"type":"number","format":"double"}}},"CategoryDetailsDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"description":{"type":"string"},"thumbnailImg":{"type":"string"}}},"UserFilterDto":{"type":"object","properties":{"username":{"type":"string"},"page":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"sortBy":{"type":"string"},"sortOrder":{"type":"string"}}},"PaginationResponseUserInAdminDto":{"type":"object","properties":{"page":{"type":"integer","description":"Current page","format":"int32"},"total":{"type":"integer","description":"Total elements","format":"int64"},"totalPages":{"type":"integer","description":"Total pages","format":"int32"},"data":{"type":"array","items":{"$ref":"#/components/schemas/UserInAdminDto"}}}},"UserInAdminDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"username":{"type":"string"},"firstname":{"type":"string"},"lastname":{"type":"string"},"createdOn":{"type":"string","format":"date-time"}}},"PaginationResponseProductCardInAdminDto":{"type":"object","properties":{"page":{"type":"integer","description":"Current page","format":"int32"},"total":{"type":"integer","description":"Total elements","format":"int64"},"totalPages":{"type":"integer","description":"Total pages","format":"int32"},"data":{"type":"array","items":{"$ref":"#/components/schemas/ProductCardInAdminDto"}}}},"ProductCardInAdminDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"thumbnailImg":{"type":"string"},"price":{"type":"integer","format":"int64"},"quantity":{"type":"integer","format":"int32"},"createdOn":{"type":"string","format":"date-time"}}},"CategoryFilterDto":{"type":"object","properties":{"name":{"type":"string"},"page":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"sortBy":{"type":"string"},"sortOrder":{"type":"string"}}},"CategoryInAdminDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"thumbnailImg":{"type":"string"},"description":{"type":"string"},"createdOn":{"type":"string","format":"date-time"}}},"PaginationResponseCategoryInAdminDto":{"type":"object","properties":{"page":{"type":"integer","description":"Current page","format":"int32"},"total":{"type":"integer","description":"Total elements","format":"int64"},"totalPages":{"type":"integer","description":"Total pages","format":"int32"},"data":{"type":"array","items":{"$ref":"#/components/schemas/CategoryInAdminDto"}}}},"ProductIdsDto":{"type":"object","properties":{"ids":{"type":"array","items":{"type":"integer","format":"int64"}}}}},"securitySchemes":{"bearerAuth":{"type":"http","name":"bearerAuth","scheme":"bearer","bearerFormat":"JWT"}}}}